feat(checkout): INT-2919 move changes to another strategy
