+++ src/payment/strategies/bolt/index.ts
@@ -1,3 +1,4 @@
 export { default as BoltPaymentStrategy } from './bolt-payment-strategy';
+export { default as BoltAppPaymentStrategy } from './bolt-app-payment-strategy';
 export * from './bolt';
 export { default as BoltScriptLoader } from './bolt-script-loader';
